// apps/backend/src/services/agents/types/types.ts
export enum AutomationTriggerType {
  TASK_OR_SUBTASK_CREATED = 'TASK_OR_SUBTASK_CREATED',
  TASK_STATUS_CHANGED = 'TASK_STATUS_CHANGED',
  TASK_ASSIGNEE_ADDED = 'TASK_ASSIGNEE_ADDED',
  TASK_ASSIGNEE_REMOVED = 'TASK_ASSIGNEE_REMOVED',
  TASK_ASSIGNEE_CHANGED = 'TASK_ASSIGNEE_CHANGED',
  TASK_DUE_DATE_ARRIVES = 'TASK_DUE_DATE_ARRIVES',
  TASK_DUE_DATE_CHANGED = 'TASK_DUE_DATE_CHANGED',
  TASK_START_DATE_ARRIVES = 'TASK_START_DATE_ARRIVES',
  TASK_START_DATE_CHANGED = 'TASK_START_DATE_CHANGED',
  TASK_PRIORITY_CHANGED = 'TASK_PRIORITY_CHANGED',
  TASK_NAME_CHANGED = 'TASK_NAME_CHANGED',
  TASK_TYPE_CHANGED = 'TASK_TYPE_CHANGED',
  TASK_LINKED = 'TASK_LINKED',
  TASK_TIME_TRACKED = 'TASK_TIME_TRACKED',
  TASK_UNBLOCKED = 'TASK_UNBLOCKED',
  CUSTOM_FIELD_CHANGED = 'CUSTOM_FIELD_CHANGED',
  TAG_ADDED = 'TAG_ADDED',
  TAG_REMOVED = 'TAG_REMOVED',
  CHECKLISTS_RESOLVED = 'CHECKLISTS_RESOLVED',
  SUBTASKS_RESOLVED = 'SUBTASKS_RESOLVED',
  EXISTING_TASK_ADDED_TO_LOCATION = 'EXISTING_TASK_ADDED_TO_LOCATION',
  MOVE_TO_LIST = 'MOVE_TO_LIST',
  DATE_BEFORE_AFTER = 'DATE_BEFORE_AFTER',
  EVERY_SCHEDULED_TIME = 'EVERY_SCHEDULED_TIME',
  CHAT_MESSAGE_POSTED = 'CHAT_MESSAGE_POSTED',
  WEBHOOK = 'WEBHOOK',
}

export enum TriggerType {
  AUTOMATION = 'AUTOMATION',
  SCHEDULED = 'SCHEDULED',
  MANUAL = 'MANUAL'
}